requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Payments
stripe==7.4.0

# Email
aiosmtplib==3.0.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
